Use StringStream to parse the nodes instead of regexes
The previous algorithm was not efficient. We were copying the whole
content into an array of lines. The new algorithm does not copy the
whole content.

We also do not use regular expressions anymore because it did not give a
lot of flexibility. The regular expressions needed to be more and more
complicated and obscure to come close to the Lorg syntax standard.

FossilOrigin-Name: 1bd9b3ede9c6dea393c13cf469d2100feda2e0f3212be9fda1fe3d251e7a627e