Write by hand the code to export the nodes to JSON
The way Haxe export things to JSON is annoying. It is inconsistent
because "it will automatically use the optimized native API if
available". For example maps are not exported as expected: the JSON
object has a "h" key containing the list of the keys (or something like
that, I just write that from what I remember).

Plus the method defines in Haxe to export JSON, we cannot control the
order of what is exported. For example we want this software to export
the `title` then the `units` then the `children` of the nodes.

Writing by hand the code to export things to JSON solves these problems.

FossilOrigin-Name: 4b129b160ed1430f74470c61d48d14f15850b70782d36435f3f551a188cc6991